+++ src/widgets/TokenWithLayout.h
@@ -40,7 +40,7 @@ signals:
 class TokenWithLayoutFactory : public TokenFactory
 {
 public:
-    virtual Token * createToken( const QString &text, const QString &iconName, int value, QWidget *parent = 0 );
+    virtual Token * createToken( const QString &text, const QString &iconName, qint64 value, QWidget *parent = 0 );
 };
 
 /**
@@ -52,7 +52,7 @@ class TokenWithLayout : public Token
 {
     Q_OBJECT
 public:
-    TokenWithLayout( const QString &text, const QString &iconName, int value, QWidget *parent = 0 );
+    TokenWithLayout( const QString &text, const QString &iconName, qint64 value, QWidget *parent = 0 );
     ~TokenWithLayout();
 
     Qt::Alignment alignment();
